+++ feature/editor/src/test/kotlin/jo/codeide/feature/editor/ExplorateurV2EditorViewModelTest.kt
@@ -93,6 +93,7 @@ class ExplorateurV2EditorViewModelTest : BaseEditorViewModelTest() {
                     tooling,
                     TestDispatcherProvider(regleMain.dispatcher),
                 ),
+            horloge = horlogeOutil,
             savedStateHandle = sauvegarde,
         )