+++ feature/editor/src/test/kotlin/jo/codeide/feature/editor/PanneauEditorViewModelTest.kt
@@ -85,6 +85,7 @@ class PanneauEditorViewModelTest : BaseEditorViewModelTest() {
                     tooling,
                     TestDispatcherProvider(regleMain.dispatcher),
                 ),
+            horloge = horlogeOutil,
             savedStateHandle = sauvegarde,
         )